22011 - move dataclass, add unit test
